extract PhoneNumberEntry to independant component
