+++ packages/ui/modals/dataPortal/PhoneEmail/fields.stories.tsx
@@ -7,7 +7,7 @@ import { getNames as getLocationNames } from '~ui/mockData/orgLocation'
 import { getNames as getServiceNames } from '~ui/mockData/orgService'
 
 import { formHookParams, PhoneEmailFormProvider, useForm } from './context'
-import { PhoneEmailFlags, PhoneNumberEntry, PhoneTypeSelect } from './fields'
+import { PhoneEmailFlags, PhoneTypeSelect } from './fields'
 
 const FormContextDecorator = (Story: StoryFn) => {
 	const form = useForm(formHookParams)
@@ -60,10 +60,6 @@ export default {
 	decorators: [FormContextDecorator],
 	excludeStories: ['phoneEmailFieldMocks'],
 } satisfies Meta<typeof Fragment>
-
-export const PhoneNumber = {
-	render: PhoneNumberEntry,
-} satisfies StoryObj<typeof PhoneNumberEntry>
 
 export const PhoneTypeSelection = {
 	render: PhoneTypeSelect,
