+++ packages/ui/modals/dataPortal/PhoneEmail/index.tsx
@@ -12,12 +12,13 @@ import { useDisclosure } from '@mantine/hooks'
 import { forwardRef } from 'react'
 
 import { Button } from '~ui/components/core/Button'
+import { PhoneNumberEntry } from '~ui/components/data-portal/PhoneNumberEntry'
 import { useSlug } from '~ui/hooks/useSlug'
 import { trpc as api } from '~ui/lib/trpcClient'
 import { ModalTitle } from '~ui/modals/ModalTitle'
 
 import { formHookParams, PhoneEmailFormProvider, useForm } from './context'
-import { PhoneEmailFlags, PhoneNumberEntry, PhoneTypeSelect } from './fields'
+import { PhoneEmailFlags, PhoneTypeSelect } from './fields'
 
 const PhoneEmailModalBody = forwardRef<HTMLButtonElement, PhoneEmailModalProps>(({ role, ...props }, ref) => {
 	const savePhone = api.orgPhone.create.useMutation()
@@ -70,7 +71,13 @@ const PhoneEmailModalBody = forwardRef<HTMLButtonElement, PhoneEmailModalProps>(
 								</>
 							) : (
 								<>
-									<PhoneNumberEntry />
+									<PhoneNumberEntry
+										phoneEntryProps={{
+											...form.getInputProps('phoneNumber'),
+											setError: (err) => form.setFieldError('phoneNumber', err),
+										}}
+										countrySelectProps={form.getInputProps('phoneCountryId')}
+									/>
 									<PhoneTypeSelect />
 								</>
 							)}
